+++ CHANGELOG.md
@@ -90,3 +90,4 @@
 * Define `.:.` composition: the same thing but with reverse order
 * Change `Nonempty` type family, family instances should be defined for each data structure
 * Define `:>` type operator for transformers
+* Define `Distributive` instance for `->` datatype